Decouple RESTEasy and Undertow
We create a new resteasy-server-common extension in charge of all the
RESTEasy heavy lifting.

The resteasy extension then just consumes the generated configuration to
initialize Undertow.
