+++ extensions/resteasy/deployment/src/main/java/io/quarkus/resteasy/deployment/ResteasyJaxrsConfigBuildItem.java
@@ -21,11 +21,11 @@ import io.quarkus.builder.item.SimpleBuildItem;
 /**
  * A build item that represents a JAX-RS config.
  */
-public final class ResteasyJaxrsConfig extends SimpleBuildItem {
+public final class ResteasyJaxrsConfigBuildItem extends SimpleBuildItem {
 
     public final String defaultPath;
 
-    public ResteasyJaxrsConfig(String defaultPath) {
+    public ResteasyJaxrsConfigBuildItem(String defaultPath) {
         this.defaultPath = defaultPath;
     }
 }
