+++ extensions/resteasy/deployment/src/main/java/io/quarkus/resteasy/deployment/ResteasyProcessor.java
@@ -1,0 +1,106 @@
+/*
+ * Copyright 2018 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.quarkus.resteasy.deployment;
+
+import java.util.Map.Entry;
+import java.util.Optional;
+
+import javax.servlet.DispatcherType;
+import javax.ws.rs.core.Application;
+
+import org.jboss.resteasy.plugins.server.servlet.HttpServlet30Dispatcher;
+
+import io.quarkus.deployment.annotations.BuildProducer;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+import io.quarkus.deployment.builditem.substrate.ReflectiveClassBuildItem;
+import io.quarkus.resteasy.common.deployment.ResteasyJaxrsProviderBuildItem;
+import io.quarkus.resteasy.runtime.ResteasyFilter;
+import io.quarkus.resteasy.runtime.RolesFilterRegistrar;
+import io.quarkus.resteasy.server.common.deployment.ResteasyInjectionReadyBuildItem;
+import io.quarkus.resteasy.server.common.deployment.ResteasyServerConfigBuildItem;
+import io.quarkus.undertow.deployment.FilterBuildItem;
+import io.quarkus.undertow.deployment.ServletBuildItem;
+import io.quarkus.undertow.deployment.ServletInitParamBuildItem;
+
+/**
+ * Processor that finds JAX-RS classes in the deployment
+ */
+public class ResteasyProcessor {
+
+    private static final String JAVAX_WS_RS_APPLICATION = Application.class.getName();
+    private static final String JAX_RS_FILTER_NAME = JAVAX_WS_RS_APPLICATION;
+    private static final String JAX_RS_SERVLET_NAME = JAVAX_WS_RS_APPLICATION;
+
+    @BuildStep
+    public void jaxrsConfig(Optional<ResteasyServerConfigBuildItem> resteasyServerConfig,
+            BuildProducer<ResteasyJaxrsConfigBuildItem> resteasyJaxrsConfig) {
+        if (resteasyServerConfig.isPresent()) {
+            resteasyJaxrsConfig.produce(new ResteasyJaxrsConfigBuildItem(resteasyServerConfig.get().getPath()));
+        }
+    }
+
+    @BuildStep
+    public void build(
+            Optional<ResteasyServerConfigBuildItem> resteasyServerConfig,
+            BuildProducer<FeatureBuildItem> feature,
+            BuildProducer<FilterBuildItem> filter,
+            BuildProducer<ServletBuildItem> servlet,
+            BuildProducer<ReflectiveClassBuildItem> reflectiveClass,
+            BuildProducer<ServletInitParamBuildItem> servletInitParameters,
+            ResteasyInjectionReadyBuildItem resteasyInjectionReady) throws Exception {
+        feature.produce(new FeatureBuildItem(FeatureBuildItem.RESTEASY));
+
+        if (resteasyServerConfig.isPresent()) {
+            String path = resteasyServerConfig.get().getPath();
+
+            //if JAX-RS is installed at the root location we use a filter, otherwise we use a Servlet and take over the whole mapped path
+            if (resteasyServerConfig.get().getPath().equals("/")) {
+                filter.produce(FilterBuildItem.builder(JAX_RS_FILTER_NAME, ResteasyFilter.class.getName()).setLoadOnStartup(1)
+                        .addFilterServletNameMapping("default", DispatcherType.REQUEST).setAsyncSupported(true)
+                        .build());
+                reflectiveClass.produce(new ReflectiveClassBuildItem(false, false, ResteasyFilter.class.getName()));
+            } else {
+                String mappingPath = getMappingPath(path);
+                servlet.produce(ServletBuildItem.builder(JAX_RS_SERVLET_NAME, HttpServlet30Dispatcher.class.getName())
+                        .setLoadOnStartup(1).addMapping(mappingPath).setAsyncSupported(true).build());
+                reflectiveClass.produce(new ReflectiveClassBuildItem(false, false, HttpServlet30Dispatcher.class.getName()));
+            }
+
+            for (Entry<String, String> initParameter : resteasyServerConfig.get().getInitParameters().entrySet()) {
+                servletInitParameters.produce(new ServletInitParamBuildItem(initParameter.getKey(), initParameter.getValue()));
+            }
+        }
+    }
+
+    /**
+     * Install the JAX-RS security provider.
+     */
+    @BuildStep
+    void setupFilter(BuildProducer<ResteasyJaxrsProviderBuildItem> providers) {
+        providers.produce(new ResteasyJaxrsProviderBuildItem(RolesFilterRegistrar.class.getName()));
+    }
+
+    private String getMappingPath(String path) {
+        String mappingPath;
+        if (path.endsWith("/")) {
+            mappingPath = path + "*";
+        } else {
+            mappingPath = path + "/*";
+        }
+        return mappingPath;
+    }
+}
