+++ extensions/smallrye-openapi/deployment/src/main/java/io/quarkus/smallrye/openapi/deployment/RESTEasyExtension.java
@@ -16,7 +16,7 @@ import org.jboss.jandex.ParameterizedType;
 import org.jboss.jandex.Type;
 
 import io.quarkus.deployment.util.ServiceUtil;
-import io.quarkus.resteasy.deployment.ResteasyJaxrsConfig;
+import io.quarkus.resteasy.deployment.ResteasyJaxrsConfigBuildItem;
 import io.smallrye.openapi.api.OpenApiConstants;
 import io.smallrye.openapi.runtime.scanner.DefaultAnnotationScannerExtension;
 import io.smallrye.openapi.runtime.scanner.OpenApiAnnotationScanner;
@@ -38,7 +38,7 @@ public class RESTEasyExtension extends DefaultAnnotationScannerExtension {
     private List<DotName> asyncTypes = new ArrayList<>();
     private String defaultPath;
 
-    public RESTEasyExtension(ResteasyJaxrsConfig jaxrsConfig, IndexView index) {
+    public RESTEasyExtension(ResteasyJaxrsConfigBuildItem jaxrsConfig, IndexView index) {
         this.defaultPath = jaxrsConfig.defaultPath;
         // the index is not enough to scan for providers because it does not contain
         // dependencies, so we have to rely on scanning the declared providers via services
